refactor: hide node-specific runtime imports from bundlers
