+++ packages/config/package.json
@@ -2,6 +2,7 @@
   "name": "@zwave-js/config",
   "version": "15.2.0",
   "description": "zwave-js: configuration files",
+  "keywords": [],
   "publishConfig": {
     "access": "public"
   },
@@ -19,7 +20,24 @@
     },
     "./package.json": "./package.json"
   },
-  "keywords": [],
+  "imports": {
+    "#default_bindings/fs": {
+      "@@dev": "@zwave-js/core/bindings/fs/node",
+      "node": "@zwave-js/core/bindings/fs/node",
+      "default": "@zwave-js/core/bindings/fs/stub"
+    },
+    "#default_bindings/log": {
+      "@@dev": "@zwave-js/core/bindings/log/node",
+      "node": "@zwave-js/core/bindings/log/node",
+      "default": "@zwave-js/core/bindings/log/stub"
+    },
+    "#config_dir": {
+      "@@dev": "./src/configDir.node.ts",
+      "browser": "./build/esm/configDir.browser.js",
+      "import": "./build/esm/configDir.node.js",
+      "require": "./build/cjs/configDir.node.js"
+    }
+  },
   "files": [
     "build/**/*.{js,cjs,mjs,d.ts,d.cts,d.mts,map}",
     "build/**/package.json",
