+++ packages/config/src/configDir.browser.ts
@@ -1,0 +1,2 @@
+/** The absolute path of the embedded configuration directory */
+export const configDir = import.meta.resolve("/config");
