+++ packages/serial/src/bindings/stub.ts
@@ -1,0 +1,4 @@
+import type { Serial } from "../serialport/Bindings.js";
+
+// Stub bindings that do nothing
+export const serial: Serial = {};
