+++ packages/zwave-js/src/lib/driver/mDNSDiscovery/stub.ts
@@ -1,0 +1,8 @@
+import type { RemoteSerialPort } from "../mDNSDiscovery.js";
+
+// On other platforns than Node.js, we cannot use mDNS discovery
+export function discoverRemoteSerialPorts(
+	_timeout: number | false = 1000,
+): Promise<RemoteSerialPort[] | undefined> {
+	return Promise.resolve([]);
+}
