+++ packages/zwave-js/src/lib/zniffer/Zniffer.ts
@@ -389,11 +389,11 @@ export class Zniffer extends TypedEventTarget<ZnifferEventCallbacks> {
 		// on Node.js internals
 		this.bindings = {
 			fs: this._options.host?.fs
-				?? (await import("@zwave-js/core/bindings/fs/node")).fs,
+				?? (await import("#default_bindings/fs")).fs,
 			serial: this._options.host?.serial
-				?? (await import("@zwave-js/serial/bindings/node")).serial,
+				?? (await import("#default_bindings/serial")).serial,
 			log: this._options.host?.log
-				?? (await import("@zwave-js/core/bindings/log/node")).log,
+				?? (await import("#default_bindings/log")).log,
 		};
 
 		// Initialize logging
